+++ packages/next/CHANGELOG.md
@@ -1,5 +1,34 @@
 # gt-next
 
+## 6.7.19
+
+### Patch Changes
+
+- Updated dependencies [[`3da05a1`](https://github.com/generaltranslation/gt/commit/3da05a12a37a62ace3c7e321aa2fed5a4af52ad9)]:
+  - generaltranslation@7.9.1
+  - @generaltranslation/compiler@1.0.9
+  - gt-react@10.6.16
+  - @generaltranslation/supported-locales@2.0.22
+
+## 6.7.18
+
+### Patch Changes
+
+- Updated dependencies [[`93881f1`](https://github.com/generaltranslation/gt/commit/93881f159455a9bbc13d14e7fec9befa60998ba3)]:
+  - generaltranslation@7.9.0
+  - @generaltranslation/compiler@1.0.8
+  - gt-react@10.6.15
+  - @generaltranslation/supported-locales@2.0.21
+
+## 6.7.17
+
+### Patch Changes
+
+- [#782](https://github.com/generaltranslation/gt/pull/782) [`155fc2c`](https://github.com/generaltranslation/gt/commit/155fc2c987078b2ffc12c55abb65bb7ff16eb09b) Thanks [@ErnestM1234](https://github.com/ErnestM1234)! - fix: only throw errors in development for invalid icu strings
+
+- Updated dependencies [[`155fc2c`](https://github.com/generaltranslation/gt/commit/155fc2c987078b2ffc12c55abb65bb7ff16eb09b)]:
+  - gt-react@10.6.14
+
 ## 6.7.16
 
 ### Patch Changes